# internal/api/assets/openapi.yaml
openapi: 3.0.3
info:
  title: Baseline API
  version: "1.0.0"
  description: "Baseline policy and dashboard API contract for implemented /v1 endpoints. Fields marked with x-baseline-implementation=reserved_noop are accepted for forward compatibility but not currently enforced."
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /v1/projects:
    get:
      summary: List projects
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
    post:
      summary: Create project
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: header
          name: X-Baseline-CSRF
          required: false
          schema:
            type: string
            enum: ["1"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ErrorResponse"
  /v1/projects/{project_id}:
    get:
      summary: Get project
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/ErrorResponse"
    put:
      summary: Update project
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
        - in: header
          name: X-Baseline-CSRF
          required: false
          schema:
            type: string
            enum: ["1"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /v1/scans:
    get:
      summary: List scans
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: query
          name: project_id
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Scan list
          content:
            application/json:
              schema:
                type: object
                properties:
                  scans:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScanSummary"
    post:
      summary: Ingest scan
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: header
          name: X-Baseline-CSRF
          required: false
          schema:
            type: string
            enum: ["1"]
        - in: header
          name: Idempotency-Key
          required: false
          description: Idempotency key for safe scan ingestion retries.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScanRequest"
      responses:
        "201":
          description: Scan accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanSummary"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
  /v1/scans/{scan_id}:
    get:
      summary: Get scan by id
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: scan_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanSummary"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /v1/scans/{scan_id}/report:
    get:
      summary: Render scan report
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: scan_id
          required: true
          schema:
            type: string
        - in: query
          name: format
          required: false
          schema:
            type: string
            enum: [json, text, sarif]
            default: json
      responses:
        "200":
          description: Scan report
          headers:
            Content-Disposition:
              description: Attachment filename for downloaded report.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  scan:
                    $ref: "#/components/schemas/ScanSummary"
            text/plain:
              schema:
                type: string
            application/sarif+json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: APIKey
    cookieAuth:
      type: apiKey
      in: cookie
      name: baseline_dashboard_session
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnauthorizedError:
      description: Unauthorized
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        repository_url:
          type: string
        default_branch:
          type: string
        policy_set:
          type: string
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        id:
          type: string
          description: Optional caller-provided identifier. If omitted, server generates one.
        name:
          type: string
        repository_url:
          type: string
          description: Optional repository metadata field used for display and audit context.
        default_branch:
          type: string
          description: Optional project default branch. Defaults to main.
        policy_set:
          type: string
          description: Optional policy set label. Defaults to baseline:prod.
    ScanViolation:
      type: object
      properties:
        policy_id:
          type: string
        severity:
          type: string
        message:
          type: string
    ScanSummary:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        commit_sha:
          type: string
        status:
          type: string
        violations:
          type: array
          items:
            $ref: "#/components/schemas/ScanViolation"
        created_at:
          type: string
          format: date-time
    CreateScanRequest:
      type: object
      required: [project_id]
      properties:
        id:
          type: string
          description: Optional caller-provided scan id. If omitted, server generates one.
        project_id:
          type: string
        commit_sha:
          type: string
        status:
          type: string
        violations:
          type: array
          items:
            $ref: "#/components/schemas/ScanViolation"
...